// src/app/models/vendor-mapping.ts
export class VendorMapping {
        public Id:any;
        public ID: any;
        public Title: string;
        public Category: string;
        public Vendor_x0020_Name: string;
        public VendorName: string;
        public Vendor_CP1: {
          Title: string;
          Name: string;
          ID: string
        };
        public Document_Location: string;
        public CD_x0020_Name: {
          Title: string;
          Name: string;
          ID: string
        };
        public CDName: {
          Title: string;
          Name: string;
          ID: string
        };
        public Discussion_x0020_Link: string;
        public DiscussionLink: string;
        public Discussion: string;
        public Vendor_CP1_Email: string;
        public Vendor_CP2_Email: string;
        public Vendor_CP3_Email: string;
        public Vendor_CP4_Email: string;
        public Vendor_CP5_Email: string;
        public OData__x0063_d1: {
          Title: string;
          Name: string;
          ID: string
        };
        public cd1: {
          Title: string;
          Name: string;
          ID: string
        };
        public OData__x0063_d2: {
          Title: string;
          Name: string;
          ID: string
        };
        public cd2: {
          Title: string;
          Name: string;
          ID: string
        };
        public OData__x0063_d3: {
          Title: string;
          Name: string;
          ID: string
        };
        public cd3: {
          Title: string;
          Name: string;
          ID: string
        };
        public OData__x0063_d4: {
          Title: string;
          Name: string;
          ID: string
        };
        public cd4: {
          Title: string;
          Name: string;
          ID: string
        };
        public Vendor_CP2: {
          Title: string;
          Name: string;
          ID: string
        };
        public Vendor_CP3: {
          Title: string;
          Name: string;
          ID: string
        };
        public Vendor_CP4: {
          Title: string;
          Name: string;
          ID: string
        };
        public Vendor_CP5: {
          Title: string;
          Name: string;
          ID: string
        };
        public Modified: any;
        public Created: any
}